requirements.txt
asgiref==3.6.0
astroid==2.14.2
colorama==0.4.6
dill==0.3.6
Django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.0.0
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.16.2
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
wrapt==1.15.0
